backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
beanie==1.23.6
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sendgrid==6.11.0
pillow==10.1.0
slowapi==0.1.9
python-slugify==8.0.1
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pymongo==4.6.0
redis==5.0.1
email-validator==2.1.0
jinja2==3.1.2 